--- MenuItem.java
@@ -1,15 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Menu;
-
-/**
- *
- * @author Jomar
- */
-public interface MenuItem {
-    public void execute();
-    
-}
