+++ src/model/Account.java
@@ -1,5 +1,10 @@
 package model;
 
+/**
+ *
+ * @author kluceycose
+ */
+
 import java.util.Date;
 
 public class Account {
