+++ src/view/ManageAccounts.java
@@ -1,8 +1,16 @@
 package view;
 
+/**
+ *
+ * @author kluceycose
+ */
+
+import Menu.Menu;
+import Menu.MenuItem;
+import Menu.MenuStack;
 import model.AccountList;
 
-public class ManageAccounts implements MenuItem{
+public class ManageAccounts implements MenuItem {
     Menu mainMenu;
     MenuStack menus;
     AccountList accountList;
