+++ test/model/TestAccounts.java
@@ -1,0 +1,5 @@
+package model;
+
+public class TestAccounts {
+
+}
