+++ test/view/TestViewAccounts.java
@@ -1,0 +1,4 @@
+package view;
+
+public class TestViewAccounts {
+}
